+++ uniffi_macros/src/export/callback_interface.rs
@@ -16,6 +16,7 @@ pub(super) fn trait_impl(
     mod_path: &str,
     trait_ident: &Ident,
     items: &[ImplItem],
+    trait_interface: bool,
 ) -> syn::Result<TokenStream> {
     let trait_name = ident_to_string(trait_ident);
     let trait_impl_ident = trait_impl_ident(&trait_name);
@@ -32,6 +33,18 @@ pub(super) fn trait_impl(
             _ => unreachable!("traits have no constructors"),
         })
         .collect::<syn::Result<TokenStream>>()?;
+
+    let uniffi_foreign_handle = trait_interface.then(|| {
+        quote! {
+            fn uniffi_foreign_handle(&self) -> Option<::uniffi::Handle> {
+                #internals_ident.invoke_callback::<(), crate::UniFfiTag>(
+                    self.handle, uniffi::IDX_CALLBACK_INC_REF, Default::default()
+                );
+                Some(self.handle)
+            }
+        }
+    });
+
     Ok(quote! {
         #[doc(hidden)]
         static #internals_ident: ::uniffi::ForeignCallbackInternals = ::uniffi::ForeignCallbackInternals::new();
@@ -66,6 +79,7 @@ pub(super) fn trait_impl(
 
         impl #trait_ident for #trait_impl_ident {
             #trait_impl_methods
+            #uniffi_foreign_handle
         }
     })
 }