+++ uniffi_macros/src/export/trait_interface.rs
@@ -4,6 +4,7 @@
 
 use proc_macro2::{Ident, Span, TokenStream};
 use quote::{quote, quote_spanned};
+use syn::ItemTrait;
 
 use crate::{
     export::{
@@ -13,7 +14,6 @@ use crate::{
     object::interface_meta_static_var,
     util::{derive_ffi_traits, ident_to_string, tagged_impl_header},
 };
-use uniffi_meta::free_fn_symbol_name;
 
 pub(super) fn gen_trait_scaffolding(
     mod_path: &str,
@@ -26,14 +26,30 @@ pub(super) fn gen_trait_scaffolding(
         return Err(syn::Error::new_spanned(rt, "not supported for traits"));
     }
     let trait_name = ident_to_string(&self_ident);
-    let trait_impl = callback_interface::trait_impl(mod_path, &self_ident, &items)
+    let trait_impl = callback_interface::trait_impl(mod_path, &self_ident, &items, true)
         .unwrap_or_else(|e| e.into_compile_error());
+    let inc_ref_fn_ident = Ident::new(
+        &uniffi_meta::inc_ref_fn_symbol_name(mod_path, &trait_name),
+        Span::call_site(),
+    );
     let free_fn_ident = Ident::new(
-        &free_fn_symbol_name(mod_path, &trait_name),
+        &uniffi_meta::free_fn_symbol_name(mod_path, &trait_name),
         Span::call_site(),
     );
 
-    let free_tokens = quote! {
+    let helper_ffi_fn_tokens = quote! {
+        #[doc(hidden)]
+        #[no_mangle]
+        pub extern "C" fn #inc_ref_fn_ident(
+            handle: ::uniffi::Handle,
+            call_status: &mut ::uniffi::RustCallStatus
+        ) {
+            uniffi::rust_call(call_status, || {
+                <dyn #self_ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::inc_ref(handle);
+                Ok(())
+            });
+        }
+
         #[doc(hidden)]
         #[no_mangle]
         pub extern "C" fn #free_fn_ident(
@@ -71,7 +87,7 @@ pub(super) fn gen_trait_scaffolding(
 
     Ok(quote_spanned! { self_ident.span() =>
         #meta_static_var
-        #free_tokens
+        #helper_ffi_fn_tokens
         #trait_impl
         #impl_tokens
         #ffi_converter_tokens
@@ -98,12 +114,26 @@ pub(crate) fn ffi_converter(mod_path: &str, trait_ident: &Ident, udl_mode: bool)
         unsafe #impl_spec {
             type FfiType = ::uniffi::Handle;
 
-            fn lower(obj: ::std::sync::Arc<Self>) -> Self::FfiType {
-                <dyn #trait_ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::insert(obj)
+            fn lower(obj: ::std::sync::Arc<Self>) -> ::uniffi::Handle {
+                // If obj wraps a foreign implementation, then `uniffi_foreign_handle` will return
+                // the handle here and we can use that rather than wrapping it again with Rust.
+                let handle = match obj.uniffi_foreign_handle() {
+                    Some(handle) => handle,
+                    None => <dyn #trait_ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::insert(obj),
+                };
+                handle
+
             }
 
-            fn try_lift(v: Self::FfiType) -> ::uniffi::deps::anyhow::Result<::std::sync::Arc<Self>> {
-                Ok(::std::sync::Arc::new(<#trait_impl_ident>::new(v)))
+            fn try_lift(handle: ::uniffi::Handle) -> ::uniffi::deps::anyhow::Result<::std::sync::Arc<Self>> {
+                Ok(if handle.is_foreign() {
+                    // For foreign handles, construct a struct that implements the trait by calling
+                    // the handle
+                    ::std::sync::Arc::new(<#trait_impl_ident>::new(handle))
+                } else {
+                    // For Rust handles, remove the `Arc<>` from our slab.
+                    <dyn #trait_ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::remove(handle)
+                })
             }
 
             fn write(obj: ::std::sync::Arc<Self>, buf: &mut Vec<u8>) {
@@ -128,6 +158,34 @@ pub(crate) fn ffi_converter(mod_path: &str, trait_ident: &Ident, udl_mode: bool)
 
         unsafe #lift_ref_impl_spec {
             type LiftType = ::std::sync::Arc<dyn #trait_ident>;
+        }
+    }
+}
+
+pub fn alter_trait(item: &ItemTrait) -> TokenStream {
+    let ItemTrait {
+        attrs,
+        vis,
+        unsafety,
+        auto_token,
+        trait_token,
+        ident,
+        generics,
+        colon_token,
+        supertraits,
+        items,
+        ..
+    } = item;
+
+    quote! {
+        #(#attrs)*
+        #vis #unsafety #auto_token #trait_token #ident #generics #colon_token #supertraits {
+            #(#items)*
+
+            #[doc(hidden)]
+            fn uniffi_foreign_handle(&self) -> Option<::uniffi::Handle> {
+                None
+            }
         }
     }
 }